+++ src/Game/Werewolf/Game.hs
@@ -175,7 +175,7 @@ stageAvailable game WitchsTurn      =
     && (not (game ^. healUsed) || not (game ^. poisonUsed))
 stageAvailable game WolfHoundsTurn  =
     has (players . wolfHounds . alive) game
-    && (isNothing $ game ^. allegianceChosen)
+    && isNothing (game ^. allegianceChosen)
 
 -- | Creates a new 'Game' with the given players. No validations are performed here, those are left
 --   to 'Game.Werewolf.Engine.startGame'.